refund-console/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.30.0
openai>=1.52.0
cachetools>=5.3.0
pydantic>=2.0.0
python-multipart>=0.0.9
requests>=2.31.0
mlflow>=2.12.0
